+++ Xext/xv/xvdisp.c
@@ -40,7 +40,7 @@ SOFTWARE.
 #include "Xext/panoramiX/panoramiX.h"
 #include "Xext/panoramiX/panoramiXsrv.h"
 #include "Xext/shm/shm_priv.h"
-#include "Xext/xvdix_priv.h"
+#include "xvdix_priv.h"
 
 #include "misc.h"
 #include "scrnintstr.h"
