+++ hw/xfree86/common/xf86xv.c
@@ -36,7 +36,7 @@
 
 #include "dix/screen_hooks_priv.h"
 #include "include/extinit.h"
-#include "Xext/xvdix_priv.h"
+#include "Xext/xv/xvdix_priv.h"
 
 #include "misc.h"
 #include "xf86.h"
